Features
Files dates are read unsuccessfully

+++ FileOrganizer/MainWindow.xaml.cs
@@ -19,6 +19,10 @@ using static System.Windows.Forms.VisualStyles.VisualStyleElement;
 using System.Windows.Media.TextFormatting;
 using System.Diagnostics;
 using System.ComponentModel;
+using System.Windows.Forms;
+using FileAway.Properties;
+using System.Globalization;
+using System;
 
 
 
@@ -38,14 +42,15 @@ namespace FileOrganizer
         public ObservableCollection<Processed> ProcessedList { get; set; }
         public DataTable excelData {  get; set; }
 
-        private Timer? timer1;
+        private System.Threading.Timer? timer1;
 
         private string gateDirectory;
 
-        private TaskbarIcon? tb;
+        private TaskbarIcon? tbi;
 
         public MainWindow()
         {
+            
             Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
             //FileList = new ObservableCollection<string>();
             FileList = new List<string>();
@@ -61,34 +66,42 @@ namespace FileOrganizer
             InitializeComponent();
             this.DataContext = this;
 
-            timer1 = new Timer(Callback, null, 0, 10000);
+            //TASKBAR ICON
+            tbi = new TaskbarIcon();
+            tbi.Icon = System.Drawing.Icon.ExtractAssociatedIcon(System.Reflection.Assembly.GetEntryAssembly().ManifestModule.Name);
+            tbi.ToolTipText = "FileAway";
+
+            timer1 = new System.Threading.Timer(Callback, null, 0, 10000);
 
             var periodicTimer = new PeriodicTimer(TimeSpan.FromSeconds(5));
-
+            
             //READ ALL .TXT FILES
             string appPath = AppContext.BaseDirectory;
             string appPathPrevious = Directory.GetParent(appPath).Parent.FullName;
             string excelPath = Path.Combine(appPathPrevious, @"data.xlsx");
-            string icoPath = Path.Combine(appPath, @"icons\Logo.ico");
-
-            //tb = (TaskbarIcon)FindResource("MyNotifyIcon");
-            //tb.Icon = new System.Drawing.Icon(icoPath);
-            TaskBar.Icon = new System.Drawing.Icon(icoPath);
-
+            
             ReadDataExcel(excelPath);
 
-            gateDirectory = Path.Combine(appPathPrevious, @"Gate");
-            //gateDirectory = @"C:\Users\Ian\Desktop\Gate";
+            gateDirectory = FileAway.Properties.Settings.Default.GateFolderPath;
+            
+            if(gateDirectory != null && Path.Exists(gateDirectory))
+            {
+                ChosenFolder.Text = "Gate Folder: " + Path.GetFileName(gateDirectory);
+                StatusMessage.Text = "Chosen Gate Folder: " + gateDirectory;
+            }
 
             string[] args = Environment.GetCommandLineArgs();
-
-            /*TaskbarIcon tbi = new TaskbarIcon();
-            tbi.Icon = Resources
-            tbi.ToolTipText = "hello world";
-            */
+            
             AddItemstoFileList(args);
 
             checkGateDirectory();
+
+            this.Closing += MainWindow_Closing;
+        }
+
+        private void MainWindow_Closing(object? sender, CancelEventArgs e)
+        {
+            tbi.Dispose();
         }
 
         public class Processed : INotifyPropertyChanged
@@ -162,12 +175,12 @@ namespace FileOrganizer
 
         private void Callback(object? state)
         {
-            checkGateDirectory();
+           checkGateDirectory();
         }
 
         private void checkGateDirectory()
         {
-            if (gateDirectory != null)
+            if (gateDirectory != null && Path.Exists(gateDirectory))
             {
                 string[] gateFiles = Directory.GetFiles(gateDirectory);
                 AddItemstoFileList(gateFiles);
@@ -197,10 +210,21 @@ namespace FileOrganizer
         {
             if (!Path.Exists(filePath))
             {
+                StatusMessage.Text = "data.xls file doesn't exist";
                 return;
             }
 
-            FileStream stream = File.Open(filePath, FileMode.Open, FileAccess.Read);
+            FileStream stream;
+
+            try
+            {
+                stream = File.Open(filePath, FileMode.Open, FileAccess.Read);
+            }
+            catch(Exception ex)
+            {
+                StatusMessage.Text = ex.Message;
+                return;
+            }
             IExcelDataReader excelReader;
 
             excelReader = ExcelDataReader.ExcelReaderFactory.CreateReader(stream);
@@ -216,17 +240,20 @@ namespace FileOrganizer
             var dataSet = excelReader.AsDataSet(conf);
 
             excelData = dataSet.Tables[0];
+
+            stream.Dispose();
         }
 
         private void AddItemstoFileList(string[] files)
         {
             foreach (string s in files)
             {
-                if (!Path.GetFileName(s).EndsWith(".dll"))
+                if (!Path.GetExtension(s).Contains(".dll"))
                 {
                     FileList.Add(s);
                 }
             }
+            
             OrganizeFiles();
         }
 
@@ -237,21 +264,70 @@ namespace FileOrganizer
                 int rowIndex = 0;
                 string fileName = Path.GetFileNameWithoutExtension(file);
                 string fullName = Path.GetFileName(file);
+                string[] fileNamePieces;
+                DateTime fileDate = DateTime.Today;
                 string? rename = null;
+                bool isDate = false;
+
+                string chosenPiece = "";
+                string stringDate = "";
 
                 if (fileName.Contains("_"))
                 {
-                    fileName = fileName.Split('_')[0];
+                    fileNamePieces = fileName.Split('_');
+                    int dateIndex = 0;
+
+                    foreach (string fileNamePiece in fileNamePieces)
+                    {
+                        //isDate = DateTime.TryParseExact(fileNamePiece, "MMddyyyy", enUS,
+                         //     DateTimeStyles.AdjustToUniversalout, fileDate);
+                        if ( isDate ) { break; }
+                        dateIndex++;
+                    }
+
+                    if (isDate)
+                    {
+                        if (dateIndex == 0)
+                        {
+                            chosenPiece = fileNamePieces[1];
+                        }
+                        else
+                        {
+                            chosenPiece = fileNamePieces[0];
+                        }
+                    }
+                    else
+                    {
+                        chosenPiece = fileNamePieces[0];
+                    }
+                }
+                else
+                {
+                    chosenPiece = fileName;
+                }
+
+
+                string finalDate = "";
+                if (isDate)
+                {
+                    stringDate = fileDate.ToShortDateString();
+                    string[] datePieces = stringDate.Split('/');
+                    finalDate = datePieces[2].Substring(2) + datePieces[1] + datePieces[0];
+                }
+                else
+                {
+                    string[] datePieces = DateTime.Today.ToShortDateString().Split('/');
+                    finalDate = datePieces[2].Substring(2) + datePieces[1] + datePieces[0];
                 }
                 
+
                 foreach (DataRow row in excelData.Rows)
                 {
                     string Keyword = row["Keyword"].ToString();
 
-                    if (fileName.Equals(Keyword))
+                    if (chosenPiece.Equals(Keyword))
                     {
                         string? filePath = null;
-                       
 
                         try
                         {
@@ -291,8 +367,16 @@ namespace FileOrganizer
                 }
 
                 string name = Path.GetFileNameWithoutExtension(file);
+                rename = finalDate + "_" + rename;
                 Processed mewItem = new Processed(name, rename);
-                Application.Current.Dispatcher.BeginInvoke(new Action(() => this.ProcessedList.Add(mewItem)));
+                try
+                {
+                    System.Windows.Application.Current.Dispatcher.BeginInvoke(new Action(() => this.ProcessedList.Add(mewItem)));
+                }
+                catch(System.Exception e)
+                {
+                    StatusMessage.Text = e.ToString();
+                }
             }
 
             ClearGateDirectory();
@@ -301,11 +385,38 @@ namespace FileOrganizer
 
         private void ClearGateDirectory()
         {
-            string[] gateFiles = Directory.GetFiles(gateDirectory);
-            foreach (string gateFile in gateFiles)
+            if(gateDirectory != null && Path.Exists(gateDirectory))
             {
-                File.Delete(gateFile);
+                string[] gateFiles = Directory.GetFiles(gateDirectory);
+                foreach (string gateFile in gateFiles)
+                {
+                    File.Delete(gateFile);
+                }
             }
+
+        }
+
+        private void GateFolder_Click(object sender, RoutedEventArgs e)
+        {
+            using (var fbd = new FolderBrowserDialog())
+            {
+                DialogResult result = fbd.ShowDialog();
+
+                string chosenFolder = fbd.SelectedPath;
+                
+                FileAway.Properties.Settings.Default.GateFolderPath = chosenFolder;
+                FileAway.Properties.Settings.Default.Save();
+
+                gateDirectory = FileAway.Properties.Settings.Default.GateFolderPath;
+
+                ChosenFolder.Text = "Gate Folder: " + Path.GetFileName(chosenFolder);
+                StatusMessage.Text = "Chosen Gate Folder: " + chosenFolder;
+            }
+        }
+
+        private void ClearProcessedList_Click(object sender, RoutedEventArgs e)
+        {
+            ProcessedList.Clear();
         }
     }
 }
